refactor: split shell model presentation
